Fixed all tests in test_proj_data_info with tilt (view_offset)
For arc corrected: Move the phi in [0,pi) after tilt. 
For No arc corrected: Shift the view_num w.r.t. the det IDs. 
At this point I have tried only the test_proj_data_info. I am not sure if the reconstruction works and moreover if it produces resonable results.

+++ src/include/stir/ProjDataInfoCylindrical.inl
@@ -67,9 +67,10 @@ initialise_ring_diff_arrays_if_not_done_yet() const
 }
 
 //PW Added view offset from the scanners. Code will now support intrinsic tilt.
+//N.E: Added proper casting.
 float
 ProjDataInfoCylindrical::get_phi(const Bin& bin)const
-{ return bin.view_num()*azimuthal_angle_sampling + scanner_ptr->get_default_intrinsic_tilt();}
+{return static_cast<float>(bin.view_num())*azimuthal_angle_sampling + scanner_ptr->get_default_intrinsic_tilt();}
 
 
 float
